Add favicon barrel Refactor folder structure
