+++ src/app/core/components/profile-search/ProfileSearchComponent.ts
