+++ src/app/profile/ProfileModule.ts
@@ -1,17 +1,16 @@
 import {THEME_MODULE} from "../theme/ThemeModule";
-import {AboutDirective} from "./components/about/AboutDirective";
+import {AboutDirective} from "./about/AboutDirective";
 import {ORDINEO_CORE} from "../core/core.module.ts";
 import {ProfileService} from "./services/ProfileService";
 import {TRAVERSON} from "../traverson/TraversonModule";
 import {ORDINEO_GATEWAY} from "../gateway/GatewayModule";
-import {TimelineComponent} from "./components/timeline/TimelineComponent";
-import {MilestoneComponent} from "./components/timeline/MilestoneComponent";
+import {TimelineComponent} from "./milestones/timeline/TimelineComponent";
+import {MilestoneComponent} from "./milestones/timeline/MilestoneComponent";
 import {MilestoneService} from "./services/MilestoneService";
-import {MilestoneContainerComponent} from "./components/timeline/MilestoneContainerComponent";
-import {MilestoneCreateComponent} from "./components/timeline/MilestoneCreateComponent";
-import {ObjectivesSearch} from "./components/objectives-search/ObjectivesSearch";
+import {MilestoneContainerComponent} from "./milestones/timeline/MilestoneContainerComponent";
+import {MilestoneCreateComponent} from "./milestones/timeline/MilestoneCreateComponent";
+import {ObjectivesSearch} from "./milestones/objectives-search/ObjectivesSearch";
 import {ProfileComponent} from "./ProfileComponent";
-import {ProfileSearchComponent} from "./components/profile-search/ProfileSearchComponent";
 
 /**
  * @ngdoc module
@@ -39,5 +38,4 @@ angular
   .component(MilestoneCreateComponent.NAME, new MilestoneCreateComponent())
   .component(ObjectivesSearch.NAME, new ObjectivesSearch())
   .component(ProfileComponent.NAME, new ProfileComponent())
-  .component(ProfileSearchComponent.NAME, new ProfileSearchComponent())
   .directive(AboutDirective.NAME, AboutDirective.instance());
