+++ src/app/profile/milestones/timeline/MilestoneComponent.ts
