+++ src/app/profile/milestones/timeline/milestone-styles.scss
