+++ src/assets/favicons/FaviconBarrel.ts
@@ -1,0 +1,18 @@
+import './android-chrome-36x36.png';
+import './android-chrome-48x48.png';
+import './android-chrome-72x72.png';
+import './android-chrome-96x96.png';
+import './android-chrome-144x144.png';
+import './android-chrome-192x192.png';
+import './apple-touch-icon-57x57.png';
+import './apple-touch-icon-60x60.png';
+import './apple-touch-icon-72x72.png';
+import './apple-touch-icon-76x76.png';
+import './apple-touch-icon-114x114.png';
+import './apple-touch-icon-120x120.png';
+import './apple-touch-icon-144x144.png';
+import './apple-touch-icon-152x152.png';
+import './apple-touch-icon-180x180.png';
+import './apple-touch-icon-precomposed.png';
+import './apple-touch-icon.png';
+import './mstile-144x144.png';
