+++ src/bootstrap.ts
@@ -1,4 +1,5 @@
 import 'angular';
 import {APP_NAME} from "./app/app.module.ts";
+import './assets/favicons/FaviconBarrel.ts';
 
 angular.bootstrap(document, [APP_NAME]);
